requirements.txt
beautifulsoup4==4.9.3
breadability==0.1.20
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
docopt==0.6.2
idna==2.10
joblib==1.0.0
lxml==4.6.2
mecab-python3==1.0.3
nltk==3.5
numpy==1.19.4
pycountry==20.7.3
regex==2020.11.13
requests==2.25.1
soupsieve==2.1
sumy==0.8.1
tinysegmenter==0.4
tqdm==4.55.0
urllib3==1.26.2
